Prevent prediction failure on zero valued variance for a given target.
The algorithm previously only predicted well if enough samples where trained
such that no zero valued variances remained for a given target.

This is a natural consequence of a strict implementation of
a gaussian based naive bayes rule, not because of a faulty implementation.

A fix is to use a negligibly small value instead of zero for the
evidence likelyhoods (based on variances) that were otherwise zero.